Moved some of the token refresh logic to the App, updated @react-oauth/google package, use the auth code flow to log in with all the cases
