+++ src/IrcMonitor.ReactUi/package.json
@@ -10,7 +10,7 @@
     "@mui/styles": "^5.11.2",
     "@mui/x-data-grid": "^5.17.21",
     "@mui/x-date-pickers": "^5.0.16",
-    "@react-oauth/google": "^0.6.0",
+    "@react-oauth/google": "^0.11.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
