+++ src/IrcMonitor.ReactUi/src/components/MenuBar.tsx
@@ -1,7 +1,7 @@
 import { AppBar, Container, IconButton, MenuItem } from "@mui/material";
-import { CredentialResponse, useGoogleLogin } from "@react-oauth/google";
-import React, { useEffect } from "react";
-import { getChannels, getIsReLogging, getUserVm, User } from "reducers/userReducer";
+import { CredentialResponse } from "@react-oauth/google";
+import React from "react";
+import { getChannels, getIsReLogging, User } from "reducers/userReducer";
 import styled from "styled-components";
 import { Menu as MenuIcon } from "@mui/icons-material";
 import { UserMenu } from "./UserMenu";
@@ -11,13 +11,10 @@ import { SelectChannel } from "./SelectChannel";
 import { userActions } from "actions/userActions";
 import { AuthorizedComponent } from "framework/AuthorizedComponent";
 import { RoleNames } from "enums/RoleEnums";
-import { useApiHook } from "hooks/useApiHook";
-import moment from "moment";
-import jwt_decode from "jwt-decode";
-import { tokenRefetchLimitInMinutes } from "framework/App";
 
 export interface MenuBarProps {
   handleGoogleAuth: (response: CredentialResponse) => void;
+  handleLoginWithGoogleAuthCode: () => void;
   handleLogOut: () => void;
   handleNavigateTo: (route: string) => void;
   user: User | undefined;
@@ -42,81 +39,19 @@ export const MenuBar: React.FC<MenuBarProps> = ({
   handleGoogleAuth,
   handleLogOut,
   user,
-  handleNavigateTo
+  handleNavigateTo,
+  handleLoginWithGoogleAuthCode
 }) => {
   // TODO Implement better handling of navigation.
 
   const channels = useSelector(getChannels);
   const isReLoggingIn = useSelector(getIsReLogging);
-  const apiHook = useApiHook();
-  const userVm = useSelector(getUserVm);
 
   const dispatch = useDispatch();
 
   const handleSelectChannel = (channelId: string | undefined) => {
     dispatch(userActions.selectChannel(channelId));
   };
-
-  useEffect(() => {
-    try {
-      // TODO handle token refresh
-      if (!userVm || !userVm.accessToken) {
-        return;
-      }
-
-      const token = jwt_decode(userVm.accessToken) as any;
-      const momentDate = moment.unix(token.exp);
-      const cur = moment();
-      var duration = moment.duration(momentDate.diff(cur)).asMinutes();
-
-      console.log(duration);
-      if (duration < tokenRefetchLimitInMinutes) {
-        if (userVm?.googleRefreshToken) {
-          dispatch(userActions.setIsLoggingIn(true));
-          apiHook.authApi
-            .authGoogleAuthCode({
-              handleGoogleAuthorizationCodeCommand: {
-                authorizationCode: userVm.googleRefreshToken,
-                email: token.sid,
-                isRefresh: true
-              }
-            })
-            .then((res) => {
-              dispatch(userActions.storeUserInfo(res));
-            })
-            .catch((err) => {
-              dispatch(userActions.setIsReLoggingIn(true));
-              console.error(err);
-            })
-            .finally(() => {
-              dispatch(userActions.setIsLoggingIn(false));
-            });
-        } else {
-          dispatch(userActions.setIsReLoggingIn(true));
-        }
-      }
-    } catch (err) {
-      console.error(err);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [apiHook.authApi, dispatch, userVm?.accessToken]);
-
-  const loginWithGoogleAuthCode = useGoogleLogin({
-    onSuccess: (code) => {
-      apiHook.authApi
-        .authGoogleAuthCode({
-          handleGoogleAuthorizationCodeCommand: { authorizationCode: code.code, email: user?.email }
-        })
-        .then((res) => {
-          dispatch(userActions.storeUserInfo(res));
-        })
-        .finally(() => {
-          dispatch(userActions.setIsReLoggingIn(false));
-        });
-    },
-    hint: user?.email,
-    flow: "auth-code"
-  });
 
   return (
     <AppBar position="fixed">
@@ -171,9 +106,7 @@ export const MenuBar: React.FC<MenuBarProps> = ({
               handleLogOut={() => {
                 handleLogOut();
               }}
-              handleGoogleAuthWithCode={() => {
-                loginWithGoogleAuthCode();
-              }}
+              handleGoogleAuthWithCode={handleLoginWithGoogleAuthCode}
               showReLogIn={isReLoggingIn}
             />
           </MenuArea>
