+++ src/IrcMonitor.ReactUi/src/framework/App.tsx
@@ -1,5 +1,5 @@
 import { Container } from "@mui/material";
-import { CredentialResponse } from "@react-oauth/google";
+import { CredentialResponse, useGoogleLogin } from "@react-oauth/google";
 import { userActions } from "actions/userActions";
 import { UserVm } from "api";
 import { MenuBar } from "components/MenuBar";
@@ -13,6 +13,8 @@ import { LocalizationProvider } from "@mui/x-date-pickers";
 import { CookiesProvider, useCookies } from "react-cookie";
 import { useApiHook } from "hooks/useApiHook";
 import { routes } from "utilities/routes";
+import jwt_decode from "jwt-decode";
+import moment from "moment";
 
 interface AppProps {
   children: React.ReactNode;
@@ -33,6 +35,7 @@ export const App: React.FC<AppProps> = (props) => {
 
   const googleTokenInfo = useSelector(getGoogleAccessToken);
 
+  // Set user info from cookies, when the app loads
   useEffect(() => {
     const userInf = cookies.userInfo as UserVm;
     if (userInf !== undefined && userInf?.email && !userVm) {
@@ -88,9 +91,72 @@ export const App: React.FC<AppProps> = (props) => {
     }
   }, [userVm, setCookie]);
 
+  // Check whether token should be refetched
+  useEffect(() => {
+    try {
+      if (!userVm || !userVm.accessToken) {
+        return;
+      }
+
+      const token = jwt_decode(userVm.accessToken) as any;
+      const momentDate = moment.unix(token.exp);
+      const cur = moment();
+      var duration = moment.duration(momentDate.diff(cur)).asMinutes();
+
+      console.log(duration);
+      if (duration < tokenRefetchLimitInMinutes) {
+        if (userVm?.googleRefreshToken) {
+          dispatch(userActions.setIsLoggingIn(true));
+          apiHook.authApi
+            .authGoogleAuthCode({
+              handleGoogleAuthorizationCodeCommand: {
+                authorizationCode: userVm.googleRefreshToken,
+                email: token.sid,
+                isRefresh: true
+              }
+            })
+            .then((res) => {
+              dispatch(userActions.storeUserInfo(res));
+            })
+            .catch((err) => {
+              dispatch(userActions.setIsReLoggingIn(true));
+              console.error(err);
+            })
+            .finally(() => {
+              dispatch(userActions.setIsLoggingIn(false));
+            });
+        } else {
+          dispatch(userActions.setIsReLoggingIn(true));
+        }
+      }
+    } catch (err) {
+      console.error(err);
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [apiHook.authApi, dispatch, userVm?.accessToken]);
+
   const handleGoogleAuth = (response: CredentialResponse) => {
     dispatch(userActions.storeGoogleAccessToken(response.credential, true));
   };
+
+  const loginWithGoogleAuthCode = useGoogleLogin({
+    onSuccess: (code) => {
+      dispatch(userActions.setIsLoggingIn(true));
+      apiHook.authApi
+        .authGoogleAuthCode({
+          handleGoogleAuthorizationCodeCommand: { authorizationCode: code.code, email: user?.email }
+        })
+        .then((res) => {
+          dispatch(userActions.storeUserInfo(res));
+        })
+        .finally(() => {
+          dispatch(userActions.setIsReLoggingIn(false));
+          dispatch(userActions.setIsLoggingIn(false));
+        });
+    },
+    hint: user?.email,
+    flow: "auth-code"
+  });
 
   const handleLogOut = () => {
     setCookie(userInfoCookieName, undefined);
@@ -111,6 +177,9 @@ export const App: React.FC<AppProps> = (props) => {
             handleGoogleAuth={handleGoogleAuth}
             handleLogOut={handleLogOut}
             handleNavigateTo={handleNavigate}
+            handleLoginWithGoogleAuthCode={() => {
+              loginWithGoogleAuthCode();
+            }}
           />
         </Container>
         <Container
